backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
